Simplify Ooo patterns. (#1)
* Pass build.

* Pass test.

* Fix test.

* Clean up.

* formatting.

* Fix gcc build.:

* FIx readme.

* Clean up comments.

* Fix win build.

* Fix win build.

* Fix win build.

* Move nbOoo == 1 out.

* Cleanup.

+++ samples/isLarge.cpp
@@ -3,7 +3,7 @@
 #include "matchit/expression.h"
 using namespace matchit;
 
-bool isLarge(double value)
+bool isLarge(int32_t value)
 {
     return match(value)(
         pattern(app([](int32_t x) { return x * x; }, _ > 1000)) = [] { return true; },
